Update Lmer.py
Pandas 2.1.0 marked DataFrame.applymap as deprecated and suggests DataFrame.map is used instead. This gives a depreciation warning when using pymer4's lmer function, and may yield an error in the future. Simply changing .applymap to .map in pymer4 would break backward compatibility, so this code checks for whether the DataFrame has the .map function (which was added in 2.1) and if so, uses it